Storage code refactor  (#987)
