+++ app/api/api/views.py
@@ -25,12 +25,12 @@ from api.serializers import (
     ScanReportViewSerializerV2,
     UserSerializer,
 )
-from azure.storage.blob import BlobServiceClient
 from config import settings
 from datasets.serializers import DataPartnerSerializer
 from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
+from django.db.models import Q
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
@@ -52,7 +52,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from shared.data.models import Concept
-from shared.files.service import delete_blob, modify_filename, upload_blob, get_blob
+from shared.jobs.models import Job, JobStage, StageStatus
 from shared.mapping.models import (
     DataDictionary,
     DataPartner,
@@ -76,8 +76,9 @@ from shared.services.rules_export import (
     get_mapping_rules_list,
     make_dag,
 )
-from shared.jobs.models import Job, JobStage, StageStatus
-from django.db.models import Q
+from shared.services.storage_service import StorageService
+
+storage_service = StorageService()
 
 
 class DataPartnerViewSet(GenericAPIView, ListModelMixin):
@@ -215,7 +216,7 @@ class ScanReportIndexV2(GenericAPIView, ListModelMixin, CreateModelMixin):
         scan_report = ScanReport.objects.create(
             dataset=valid_dataset,
             parent_dataset=valid_parent_dataset,
-            name=modify_filename(valid_scan_report_file, dt, rand),
+            name=storage_service.modify_filename(valid_scan_report_file, dt, rand),
             visibility=valid_visibility,
         )
 
@@ -230,6 +231,11 @@ class ScanReportIndexV2(GenericAPIView, ListModelMixin, CreateModelMixin):
         if sr_editors := valid_editors:
             scan_report.editors.add(*sr_editors)
 
+        # Spreadsheet Content Type
+        spreadsheet_content_type = (
+            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
+        )
+
         # If there's no data dictionary supplied, only upload the scan report
         # Set data_dictionary_blob in Azure message to None
         if str(valid_data_dictionary_file) == "None":
@@ -239,11 +245,12 @@ class ScanReportIndexV2(GenericAPIView, ListModelMixin, CreateModelMixin):
                 "data_dictionary_blob": "None",
             }
 
-            upload_blob(
+            storage_service.upload_blob(
                 scan_report.name,
                 "scan-reports",
                 valid_scan_report_file,
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+                spreadsheet_content_type,
+                use_read_method=False,
             )
 
         else:
@@ -261,17 +268,19 @@ class ScanReportIndexV2(GenericAPIView, ListModelMixin, CreateModelMixin):
                 "data_dictionary_blob": data_dictionary.name,
             }
 
-            upload_blob(
+            storage_service.upload_blob(
                 scan_report.name,
                 "scan-reports",
                 valid_scan_report_file,
-                "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+                spreadsheet_content_type,
+                use_read_method=False,
             )
-            upload_blob(
+            storage_service.upload_blob(
                 data_dictionary.name,
                 "data-dictionaries",
                 valid_data_dictionary_file,
                 "text/csv",
+                use_read_method=False,
             )
 
         # send to the upload queue
@@ -308,12 +317,14 @@ class ScanReportDetailV2(
 
     def perform_destroy(self, instance):
         try:
-            delete_blob(instance.name, "scan-reports")
+            storage_service.delete_blob(instance.name, "scan-reports")
         except Exception as e:
             raise Exception(f"Error deleting scan report: {e}")
         if instance.data_dictionary:
             try:
-                delete_blob(instance.data_dictionary.name, "data-dictionaries")
+                storage_service.delete_blob(
+                    instance.data_dictionary.name, "data-dictionaries"
+                )
             except Exception as e:
                 raise Exception(f"Error deleting data dictionary: {e}")
         instance.delete()
@@ -831,7 +842,7 @@ class DownloadScanReportViewSet(viewsets.ViewSet):
         # TODO: This should not be a list view...
         scan_report = ScanReport.objects.get(id=pk)
         blob_name = scan_report.name
-        scan_report_blob = get_blob(blob_name, "scan-reports")
+        scan_report_blob = storage_service.get_blob(blob_name, "scan-reports")
 
         response = HttpResponse(
             scan_report_blob,
