+++ app/shared/shared/services/utils.py
@@ -1,0 +1,85 @@
+from typing import Any, Dict, List
+
+
+def remove_BOM(intermediate: List[Dict[str, Any]]):
+    """
+    Given a list of dictionaries, remove any occurrences of the BOM in the keys.
+
+    Args:
+        intermediate (List[Dict[str, Any]]): List of dictionaries to remove from.
+
+    Returns:
+        The list of dictionaries with BOM removed from the keys.
+    """
+    return [
+        {key.replace("\ufeff", ""): value for key, value in d.items()}
+        for d in intermediate
+    ]
+
+
+def process_three_item_dict(three_item_data):
+    """
+    Converts a list of dictionaries (each with keys 'csv_file_name', 'field_name' and
+    'code') to a nested dictionary with indices 'csv_file_name', 'field_name' and
+    internal value 'code'.
+
+    [{'csv_file_name': 'table1', 'field_name': 'field1', 'value': 'value1', 'code':
+    'code1'},
+    {'csv_file_name': 'table1', 'field_name': 'field2', 'value': 'value2'},
+    {'csv_file_name': 'table2', 'field_name': 'field2', 'value': 'value2', 'code':
+    'code2'},
+    {'csv_file_name': 'table3', 'field_name': 'field3', 'value': 'value3', 'code':
+    'code3'}]
+    ->
+    {'table1': {'field1': 'value1', 'field2': 'value2'},
+    'table2': {'field2': 'value2'},
+    'table3': {'field3': 'value3}
+    }
+    """
+    csv_file_names = set(row["csv_file_name"] for row in three_item_data)
+
+    # Initialise the dictionary with the keys, and each value set to a blank dict()
+    new_vocab_dictionary = {filename: {} for filename in csv_file_names}
+
+    # Fill each subdict with the data from the input list
+    for row in three_item_data:
+        new_vocab_dictionary[row["csv_file_name"]][row["field_name"]] = row["code"]
+
+    return new_vocab_dictionary
+
+
+def process_four_item_dict(four_item_data):
+    """
+    Converts a list of dictionaries (each with keys 'csv_file_name', 'field_name' and
+    'code' and 'value') to a nested dictionary with indices 'csv_file_name',
+    'field_name', 'code', and internal value 'value'.
+
+    [{'csv_file_name': 'table1', 'field_name': 'field1', 'value': 'value1', 'code':
+    'code1'},
+    {'csv_file_name': 'table1', 'field_name': 'field2', 'value': 'value2', 'code':
+    'code2'},
+    {'csv_file_name': 'table2', 'field_name': 'field2', 'value': 'value2', 'code':
+    'code2'},
+    {'csv_file_name': 'table2', 'field_name': 'field2', 'value': 'value3', 'code':
+    'code3'},
+    {'csv_file_name': 'table3', 'field_name': 'field3', 'value': 'value3', 'code':
+    'code3'}]
+    ->
+    {'table1': {'field1': {'value1': 'code1'}, 'field2': {'value2': 'code2'}},
+    'table2': {'field2': {'value2': 'code2', 'value3': 'code3'}},
+    'table3': {'field3': {'value3': 'code3'}}
+    }
+    """
+    csv_file_names = set(row["csv_file_name"] for row in four_item_data)
+
+    # Initialise the dictionary with the keys, and each value set to a blank dict()
+    new_data_dictionary = dict.fromkeys(csv_file_names, {})
+
+    for row in four_item_data:
+        if row["field_name"] not in new_data_dictionary[row["csv_file_name"]]:
+            new_data_dictionary[row["csv_file_name"]][row["field_name"]] = {}
+        new_data_dictionary[row["csv_file_name"]][row["field_name"]][row["code"]] = row[
+            "value"
+        ]
+
+    return new_data_dictionary
