+++ app/workers/RulesFileQueue/__init__.py
@@ -15,18 +15,16 @@ django.setup()
 
 from django.db.models.query import QuerySet
 from shared.files.models import FileDownload, FileType
-from shared.files.service import upload_blob_read
 from shared.mapping.models import MappingRule, ScanReport
 from shared.services.rules_export import (
     get_mapping_rules_as_csv,
     get_mapping_rules_json,
     make_dag,
 )
-from shared_code.db import (
-    update_job,
-    JobStageType,
-    StageStatusType,
-)
+from shared.services.storage_service import StorageService
+from shared_code.db import JobStageType, StageStatusType, update_job
+
+storage_service = StorageService()
 
 
 def create_json_rules(rules: QuerySet[MappingRule]) -> BytesIO:
@@ -126,7 +124,9 @@ def main(msg: func.QueueMessage) -> None:
 
     # Save to blob
     filename = f"Rules - {scan_report.dataset} - {scan_report_id} - {datetime.now()}.{file_extension}"
-    upload_blob_read(filename, "rules-exports", file, file_type)
+    storage_service.upload_blob(
+        filename, "rules-exports", file, file_type, use_read_method=True
+    )
 
     # create entity
     file_type_entity = FileType.objects.get(value=file_type_value)
