+++ docker-compose.yml
@@ -70,6 +70,7 @@ services:
       - WORKERS_RULES_EXPORT_NAME=rules-exports-local
       - STORAGE_CONN_STRING=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://azurite:10000/devstoreaccount1;QueueEndpoint=http://azurite:10001/devstoreaccount1;TableEndpoint=http://azurite:10002/devstoreaccount1;
       - SIGNING_KEY=secret
+      - STORAGE_TYPE=azure
     volumes:
       - ./app/api:/api
       - ./app/shared:/shared
@@ -106,6 +107,7 @@ services:
       - DB_NAME=postgres
       - DB_USER=postgres
       - DB_PASSWORD=postgres
+      - STORAGE_TYPE=azure
       - PYTHONPATH=/home/site/wwwroot:/home/site/wwwroot/shared
     volumes:
       - ./app/workers:/home/site/wwwroot
